requirements.txt
discord.py
telethon
irc
python-dotenv